Filter for valid authorities when checking an access token for access.
Do not allow access to users that only have authorities that are not
recognised by TranSMART.
